Implemented AdgPair as "special" type
lgob relies on the underline API to provide getter/setter methods.
Although it works fine with GObject items, this approach is not enough
for GBoxed types and structs because, usually, the fields are expected
to be accessed directly from the application (after all, they are C
libraries).

This patch explores a way to solve the problem without touching lgob:
it considers AdgPair (a GBoxed type) as a "special" type and provides
__index and __newindex implementation for accessing the fields.