Implemented AdgMatrix as special type
Provided a way to access cairo matrix fields (AdgMatrix is an alias for
cairo_matrix_t) in a similar way of what done for AdgPair.

Furthermore, connected the base types (specifically GObject and
GInitiallyUnowned) to the GObject bindings so the builtinproperty
setters and getters provided by lgob now work as expected.