[build] Include Lua parsing result from different directory
Previously the parsing operation resulted in the generation of a
module.def file that was expected to be in the same directory of
module.ovr, as shown by the include() helper function in utils.lua.

In VPATH builds the generated files are in a different directory than
that of the source files, so another approach should be used. Instead of
include()ing the def file, now the parsing operation generates api.lua
that is included by module.ovr throught the standard require()
mechanism, hence allowing the specification of different inclusion paths
with LUA_PATH.

The makefiles have been modified to specify $(abs_builddir) and
$(abs_srcdir) as alternative paths.